Implemented package installation
Now it's possible to install package with bpm install.

NOTE: When you install some package that is already installeed (inside
deps dir), it will be RESETED to the origin/main state. It can override
customizations/changes in the local repo. But it will not occurs for any
local repository, only remote repos, installed from install command.