Added confirm dialog to remove local packages
When do you try to remove packages outside deps dir it will ask to
confirm. It's necessary because deps dir is used for remote packages,
that can be reinstalled, so it doesn't need to be protected.
But it isn't the case of packages defined outside dedps dir, that can be
locally dedfined packages.