Fixed sudo problems
After tested on a Ubuntu device, I've found a lot of problems related to
sudo permissions while running install command.

The way I've found to solve these problems is to block the install
script to be runned with sudo, and calling sudo -v inside the script to
ask for permission. Once the credentials is provided, the script can
access its permission without asking for the password multiple times.

NOTE: The only thing that uses sudo is the dependency installtion
process (used to ensure installed git, curl, etc)