Add "daemon" script (Unstable)
It's not using systemd, but is planned to do. It's actually just a background script,
and the reason to make this, is to reduce the time to run bpm commands, by moving the
locator update to a background script... It can also be used to scan the
entire file system for unindexed pacakges, but it might be better
planned in order to implement it.

Actually, I'll not merge it into the main branch, due to the
instabilities, and cause it appears to be more like a workaround than a
fully functional feature.